requirements.txt
Flask==2.3.3
playwright==1.40.0
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.1.0
